+++ spec/Tree/RecordSpec.php
@@ -12,7 +12,7 @@ class RecordSpec extends ObjectBehavior
 {
     function let()
     {
-        $this->beConstructedWith(0);
+        $this->beConstructedWith(0, '');
     }
 
     function it_is_initializable()
@@ -22,7 +22,8 @@ class RecordSpec extends ObjectBehavior
 
     function it_contains_a_name()
     {
-        $this->getName()->shouldEqual('Record');
+        $this->beConstructedWith(0, 'custom-name');
+        $this->getName()->shouldEqual('custom-name');
     }
 
     function it_contains_a_type()
@@ -32,7 +33,7 @@ class RecordSpec extends ObjectBehavior
 
     function it_contains_a_line_number()
     {
-        $this->beConstructedWith(123);
+        $this->beConstructedWith(123, '');
         $this->getLineNr()->shouldEqual(123);
     }
 
@@ -40,7 +41,7 @@ class RecordSpec extends ObjectBehavior
     {
         $node1->getName()->willReturn('node1');
         $node2->getName()->willReturn('node2');
-        $this->beConstructedWith(0, $node1, $node2);
+        $this->beConstructedWith(0, '', $node1, $node2);
         $this->getChild('node1')->shouldEqual($node1);
         $this->getChild('node2')->shouldEqual($node2);
     }