+++ src/Visitor/Visitor.php
@@ -28,9 +28,7 @@ use byrokrat\autogiro\Tree\Node;
  * Visitor that dynamically calls method based on node name and type
  *
  * Will read the node name and type (eg. SomeNode) and dispatch a node specific
- * method if defined in visitor (eg. beforeSomeNode or afterSomeNode). By
- * convention such visitor methods should type hint the specific node type
- * (eg. beforeSomeNode(SomeNode $node)).
+ * method if defined in visitor (eg. beforeSomeNode or afterSomeNode).
  */
 class Visitor implements VisitorInterface
 {